feat: Added Keyword Controller
